% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoading.R
\name{importGraspableJson}
\alias{importGraspableJson}
\title{importGraspableJson}
\usage{
importGraspableJson(filename, singleVersionOnly = F, limitExperiment = F)
}
\arguments{
\item{fileName}{a vector of stimuli, between 0 and inf
Comments}
}
\value{
A list of Tibbles
}
\description{
importGraspableJson
}
\examples{
canvasSnippetFile <- system.file("extdata", "canvasSnippet.json", package = "GraspableDataAnalysis", mustWork = TRUE)
canvasSnippetData <- importGraspableJson(canvasSnippetFile, singleVersionOnly = "2.5.12" )

}
